Merged PR 25: Add subscribe unsubscribe actions on venue landing page.
When loading the venue landing page first check if the user has subscribed to this venue by making a GET call /subscribe-venue/?parent_id=123&venue_id=456. If user has subscribed then display a green icon(to be provided). If user has not subscribed display a grey icon(to be provided).

1) If user clicks on grey icon do POST on /subscribe-venue to add subscription and change the icon to green.
2) If user clicks on green icon do DELETE on /subscribe-venue to delete subscription and change the icon to grey